sortby to module
