+++ polars/polars-lazy/src/physical_plan/expressions/mod.rs
@@ -4,6 +4,7 @@ pub mod default;
 mod final_agg;
 pub(crate) mod literal;
 pub(crate) mod sort;
+pub(crate) mod sortby;
 pub(crate) mod take;
 pub(crate) mod window;
